+++ src/main/java/org/tingyu/atomicarchitect/common/world/inventory/MatterDecomposerMenu.java
@@ -1,0 +1,53 @@
+package org.tingyu.atomicarchitect.common.world.inventory;
+
+import net.minecraft.world.SimpleContainer;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.inventory.*;
+
+public class MatterDecomposerMenu extends AbstractContainerMenu {
+    public static final int RESULT_SLOT = 0;
+    private static final int CRAFT_SLOT_START = 1;
+    private static final int CRAFT_SLOT_END = 10;
+    private static final int INV_SLOT_START = 10;
+    private static final int INV_SLOT_END = 37;
+    private static final int USE_ROW_SLOT_START = 37;
+    private static final int USE_ROW_SLOT_END = 46;
+    private final CraftingContainer craftSlots = new CraftingContainer(this, 1, 1);
+    private final SimpleContainer resultSlots = new SimpleContainer(9);
+    private final ContainerLevelAccess access;
+    private final Player player;
+
+    public MatterDecomposerMenu(int containerId, Inventory inventory) {
+        this(containerId, inventory, ContainerLevelAccess.NULL);
+    }
+
+    public MatterDecomposerMenu(int containerId, Inventory inventory, ContainerLevelAccess access) {
+        super(ModMenuTypes.MATTER_DECOMPOSER, containerId);
+        this.access = access;
+        this.player = inventory.player;
+        this.addSlot(new ResultSlot(inventory.player, this.craftSlots, this.resultSlots, 0, 124, 35));
+
+        for(int i = 0; i < 3; ++i) {
+            for(int j = 0; j < 3; ++j) {
+                this.addSlot(new Slot(this.craftSlots, j + i * 3, 30 + j * 18, 17 + i * 18));
+            }
+        }
+
+        for(int k = 0; k < 3; ++k) {
+            for(int i1 = 0; i1 < 9; ++i1) {
+                this.addSlot(new Slot(inventory, i1 + k * 9 + 9, 8 + i1 * 18, 84 + k * 18));
+            }
+        }
+
+        for(int l = 0; l < 9; ++l) {
+            this.addSlot(new Slot(inventory, l, 8 + l * 18, 142));
+        }
+    }
+
+
+    @Override
+    public boolean stillValid(Player p_38874_) {
+        return false;
+    }
+}