+++ src/main/java/org/tingyu/atomicarchitect/common/world/inventory/ModMenuTypes.java
@@ -1,0 +1,10 @@
+package org.tingyu.atomicarchitect.common.world.inventory;
+
+import net.minecraft.world.inventory.MenuType;
+import net.minecraftforge.registries.ObjectHolder;
+import org.tingyu.atomicarchitect.AtomicArchitect;
+
+@ObjectHolder(AtomicArchitect.MOD_ID)
+public class ModMenuTypes {
+    public static final MenuType<MatterDecomposerMenu> MATTER_DECOMPOSER = null;
+}