+++ src/main/java/org/tingyu/atomicarchitect/common/world/item/ElementItem.java
@@ -1,47 +1,41 @@
 package org.tingyu.atomicarchitect.common.world.item;
 
-import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
 
-import java.util.List;
-
-public class AtomElement extends Item {
+public class ElementItem extends Item {
     private final int atomicNumber;
     private final String name;
     private final String symbol;
-    public AtomElement(AtomElement.Properties properties) {
+
+    public ElementItem(ElementItem.Properties properties) {
         super(properties);
         this.atomicNumber = properties.atomicNumber;
         this.name = properties.name;
         this.symbol = properties.symbol;
     }
 
-    public static class Properties extends Item.Properties{
-        public Properties() {
-            this.tab(null);
-        }
-
+    public static class Properties extends Item.Properties {
         int atomicNumber;
         String name;
         String symbol;
-        public AtomElement.Properties atomicNumber(int atomicNumber) {
+
+        public Properties() {
+            super.tab(ModCreativeModeTab.TAB_ATOMIC_ARCHITECT);
+        }
+
+        public ElementItem.Properties atomicNumber(int atomicNumber) {
             this.atomicNumber = atomicNumber;
             return this;
         }
 
-        public AtomElement.Properties name(String name) {
+        public ElementItem.Properties name(String name) {
             this.name = name;
             return this;
         }
 
-        public AtomElement.Properties symbol(String symbol) {
+        public ElementItem.Properties symbol(String symbol) {
             this.symbol = symbol;
             return this;
-        }
-
-        @Override
-        public Item.Properties tab(CreativeModeTab tab) {
-            return super.tab(ModCreativeModeTab.TAB_ATOMIC_ARCHITECT);
         }
     }
 }